Project
phone no int to Long Correction

+++ Hotel Management System/Hotel-Management-System/src/main/java/com/example/Hotel/Management/System/Entity/Hotel.java
@@ -59,7 +59,7 @@ public class Hotel {
         return phoneNo;
     }
 
-    public void setPhoneNo(int phoneNo) {
+    public void setPhoneNo(long phoneNo) {
         this.phoneNo = phoneNo;
     }
